Refactor website and homepage style to fix layout
The website's general structure was refactored to fix an issue
where page contents were overflowing into the left side of the
viewport and were not visible.

The homepage was refactored to fix an issue where overflowing
content from the hero message was covering the header/ nav content.
It was also performed to prevent the layout from breaking at small
resolutions.