Refactor code for content overflow issues
This commit fixes the following bugs: (1) unviewable content on left side
of page and (2) background/overlay not expanding with overflow content.

Adding a minimum heights and widths to the css classes allows parent
containers to grow with overflowing child elements as needed and prevent
the page content from overflowing to the left side of the viewport.

Auto margins were also added due to an inherent behavior for flex box to
"truely" center an element (elements will remained centered even with
overflow). Auto margins prevent the overflow content from moving into
unviewable parts of the viewport.

The background/overlay was shifted to also help fix the second bug.