Fix, restyle, and refactor contact page
This commit fixes a bug where the navigation section of page centers
itself as the viewport changes but not the main content. A slight refactor
of the html and restyle was also performed for the main content.

The min width of the root psuedo class was increased to prevent the
the navigation from moving/ centering without the main content of the
contact page.

The content page was refactored to remove an unneeded container for the
google maps iframe.

The restyle was to adjust the spacing and size of the contact form and
business information.